+++ src/main/java/kr/where/backend/member/Member.java
@@ -1,12 +1,10 @@
 package kr.where.backend.member;
 
 import jakarta.persistence.*;
-import kr.where.backend.api.mappingDto.CadetPrivacy;
-import kr.where.backend.exception.request.RequestErrorCode;
-import kr.where.backend.exception.request.RequestException;
+import kr.where.backend.api.json.CadetPrivacy;
 import kr.where.backend.group.entity.GroupMember;
 import kr.where.backend.location.Location;
-import kr.where.backend.api.mappingDto.Hane;
+import kr.where.backend.api.json.Hane;
 import lombok.*;
 import lombok.extern.slf4j.Slf4j;
 
@@ -29,7 +27,7 @@ public class Member {
     private Long id;
 
     @Column(name = "intra_id", unique = true)
-    private Long intraId;
+    private Integer intraId;
 
     @Column(length = 15, unique = true, nullable = false)
     private String intraName;
